Added Kraken websocket book checksum.
████ ███  To request new features or in case this commit breaks something for you,
████ ███  please, create a new github issue with all possible information for me,
▓███▀█▄   but never share your API Keys!
▒▓██ ███
 _______________________________
/ Hello, WORLD!                 \
|                               |
\ pssst.. 1 BTC = 38591.16 EUR. /
 -------------------------------
        \   ^__^
         \  (oo)\_______
            (__)\       )\/\
                ||----w |
                ||     ||

+++ src/lib/Krypto.ninja-data.h
@@ -704,7 +704,7 @@ namespace ₿ {
         transform(input.begin(), input.end(), input.begin(), ::toupper);
         return input;
       };
-      static int CRC32(const string &input) {
+      static unsigned long CRC32(const string &input) {
         return crc32(0, (const Bytef*)input.data(), input.length());
       };
       static string B64(const string &input) {
